# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Indian Postal Department Field Operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /offices:
    get:
      operationId: listOffices
      tags: [offices]
      summary: List all post offices
      security:
        - bearerAuth: []
      parameters:
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of offices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Office"
    post:
      operationId: createOffice
      tags: [offices]
      summary: Create a new post office (super admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOfficeRequest"
      responses:
        "201":
          description: Office created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Office"
  /offices/{id}:
    get:
      operationId: getOffice
      tags: [offices]
      summary: Get a specific office
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Office details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Office"
    put:
      operationId: updateOffice
      tags: [offices]
      summary: Update office details or polygon
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOfficeRequest"
      responses:
        "200":
          description: Updated office
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Office"
  /offices/{id}/status:
    patch:
      operationId: setOfficeStatus
      tags: [offices]
      summary: Activate or deactivate an office
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Updated office
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Office"

  # ── Users ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StatusUpdate:
      type: object
      properties:
        isActive:
          type: boolean
      required: [isActive]
...
    Office:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
        address:
          type: string
        district:
          type: string
        state:
          type: string
        pincode:
          type: string
        phone:
          type: string
          nullable: true
        locationLat:
          type: string
          nullable: true
          description: Pin marking where the office physically is (distinct from polygonGeoJson, the area it serves)
        locationLng:
          type: string
          nullable: true
        polygonGeoJson:
          type: object
          nullable: true
          description: GeoJSON polygon of office boundary
        isActive:
          type: boolean
        operatorCount:
          type: integer
        beatCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, code, address, district, state, pincode, isActive, operatorCount, beatCount, createdAt]
    CreateOfficeRequest:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        address:
          type: string
        district:
          type: string
        state:
          type: string
        pincode:
          type: string
        phone:
          type: string
        locationLat:
          type: string
        locationLng:
          type: string
        polygonGeoJson:
          type: object
      required: [name, code, address, district, state, pincode]
    UpdateOfficeRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        district:
          type: string
        state:
          type: string
        pincode:
          type: string
        phone:
          type: string
        locationLat:
          type: string
        locationLng:
          type: string
        polygonGeoJson:
          type: object

    # ── Beat ──────────────────────────────────────────────────────────
...